requirements.txt
# Core Dependencies
beautifulsoup4==4.12.2
requests==2.31.0
fake-useragent==1.5.1
diskcache==5.6.3

# NLP & Machine Learning
transformers==4.35.2
torch==2.0.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
faiss-cpu==1.7.4

# Utilities
lxml==4.9.3
tqdm==4.66.1
logging==0.4.9.6
